backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
weasyprint==60.2
openai==1.3.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
markdown==3.5.1
spacy==3.7.2
nltk==3.8.1 